% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks}
\alias{breaks}
\title{Breaks values}
\usage{
breaks(x)
}
\arguments{
\item{x}{an output of the \code{base::cut} function.}
}
\value{
a numeric vector of break values.
}
\description{
Returns the breaks values of the output of the \code{base::cut} function
}
\examples{
x <- rnorm(10000)
y <- cut(x, breaks = -6:6)
breaks(y)

}
\author{
Marc Choisy
}
