% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Data Sets}
\usage{
datasets(pkg)
}
\arguments{
\item{pkg}{the name of a package}
}
\description{
Shows the datasets of a package
}
\details{
Returns a character vector containing the names of the datasets in the
package given as the argument
}
