% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{Fix names of provinces}
\usage{
fix_names(x)
}
\arguments{
\item{x}{a vector of characters.}
}
\value{
A vector of characters, the corrected version of the x input.
}
\description{
\code{fix_names} allows to correct the names of the provinces, adopting a
standard convention.
}
\details{
Ease the integration of data from multiple sources.
}
\examples{
fix_names(c("Soc T rang", "SOn ia", "Son ia"))
}
\author{
Marc Choisy
}
