% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_seq2.R
\name{full_seq2}
\alias{full_seq2}
\title{Create the full sequence of values in a vector}
\usage{
full_seq2(x, period, tol = 1e-06)
}
\arguments{
\item{x}{A numeric vector.}

\item{period}{Gap between each observation. The existing data will be checked
to ensure that it is actually of this periodicity. Automatically
calculated if missing.}

\item{tol}{Numerical tolerance for checking periodicity.}
}
\description{
This function is a wrapper around the \code{\link[tidyr]{full_seq}} function
and automatically calculates the smallest period is missing.
}
\seealso{
\code{\link[tidyr]{full_seq}}
}
\author{
Marc Choisy
}
