% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2vec.R
\name{mat2vec}
\alias{mat2vec}
\title{Converts matrix to vector}
\usage{
mat2vec(mat, margin = 2)
}
\arguments{
\item{mat}{the matrix to convert to a vector}

\item{margin}{indicates whether NA values separate the rows (margin = 1) or
the columns (margin = 2). Default value is 2}
}
\description{
\code{mat2vec} converts a matrix to a NA-separated vector where the NA
values separate either rows or columns.
}
\author{
Marc Choisy
}
