% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcutils.R
\docType{package}
\name{mcutils}
\alias{mcutils}
\alias{mcutils-package}
\title{mcutils: A package for complementing the base utils package.}
\description{
mcutils: A package for complementing the base utils package.
}
