% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovv.R
\name{ovv}
\alias{ovv}
\title{Overview of a data frame.}
\usage{
ovv(x, n = 6L, digits = 4L, interspace = 3L)
}
\arguments{
\item{x}{a data frame, matrix or table.}

\item{digits}{integer indicating the number of decimal places (round) or
significant digits (signif) to be used (default value: 4). See
the \code{base::round} function.}

\item{interspace}{integer indicating the number of dots to print between the
head and the tail of the object.}
}
\value{
Invisibly return the input data frame.
}
\description{
\code{ovv} prints an overview of a data frame.
}
\details{
Mostly used for in side effect printing of an overview of the data frame.
}
\examples{
# overviews of data.frames:
ovv(mtcars)
ovv(iris)

# overviews of matrices:
ovv(as.matrix(mtcars))
ovv(as.matrix(iris))

# overviews of tables:
library(magrittr) # for \%>\%
library(lubridate) # for ymd, isoweek, isoyear
# a fonction used to create the data set below:
fct <- function(y) {
g <- function(x, y) ymd(as.integer(paste0(y, x)))
  start_end <- lapply(c("0101", "1231"), g, y)
  seq(start_end[[1]], start_end[[2]], by = "day")
}
# creating a fake data set:
cal <- 10:17 \%>\%
  lapply(fct) \%>\%
  do.call(c, .) \%>\%
  (function(x) lapply(list(isoweek, isoyear), function(f) f(x))) \%>\%
  do.call(cbind, .) \%>\%
  as.data.frame \%>\%
  setNames(c("week", "year")) \%>\%
  unique
# an overview of its table:
ovv(table(cal))

}
\author{
Marc Choisy
}
