% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs2.R
\name{pairs2}
\alias{pairs2}
\title{Scatterplot matrices}
\source{
href{https://stackoverflow.com/questions/22810309/pairs-specifying-axes-limits-of-the-subpanels}{Specifying axes limits of the subpanels}
        from \href{https://stackoverflow.com}{StackOverflow}.
}
\usage{
pairs2(x, labels, panel = points, ..., lower.panel = panel,
  upper.panel = panel, diag.panel = NULL, text.panel = textPanel,
  label.pos = 0.5 + has.diag/3, line.main = 3, cex.labels = NULL,
  font.labels = 1, row1attop = TRUE, gap = 1, log = "", xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{x}{the coordinates of points given as numeric columns of a matrix or
data frame. Logical and factor columns are converted to numeric in
the same way that \code{\link[base]{data.matrix}} does.}

\item{labels}{the names of the variables.}

\item{panel}{\code{function(x, y, ...)} which is used to plot the contents of
each panel of the display.}

\item{...}{arguments to be passed to or from methods. Also,
\code{\link[graphics]{par}} parameters can be given as can
arguments to plot such as \code{main}. \code{par("oma")} will be
set appropriately unless specified.}

\item{lower.panel, upper.panel}{separate panel functions (or \code{NULL}) to
be used below and above the diagonal respectively.}

\item{diag.panel}{optional \code{function(x, ...)} to be applied on the diagonals.}

\item{text.panel}{optional \code{function(x, y, labels, cex, font, ...)} to
be applied on the}

\item{label.pos}{\code{y} position of labels in the text panel.}

\item{line.main}{if \code{main} is specified, \code{line.main} gives the line
argument to \code{\link[graphics]{mtext}}() which draws the
title. You may want to specify \code{oma} when changing
\code{line.main}.}

\item{cex.labels, font.labels}{graphics parameters for the text panel.}

\item{row1attop}{logical. Should the layout be matrix-like with row 1 at the
top, or graph-like with row 1 at the bottom?}

\item{gap}{distance between subplots, in margin lines.}

\item{log}{a character string indicating if logarithmic axes are to be used:
see \code{\link[graphics]{plot.default}}. \code{log = "xy"}
specifies logarithmic axes for all variables.}

\item{xlim, ylim}{3-dimension array. The first and second dimensions correspond
to the row and column indexes of the subplots respectively
and the third dimension is equal to 2 and correspond to the
first and second values of the \code{xlim} and \code{ylim}
arguments to use for each subplot.}
}
\description{
A matrix of scatterplots is produced
}
\details{
This function is a slight modification of the original
\code{\link[graphics]{pairs}} function as proposed here by
\href{https://stackoverflow.com/users/2591234/shadow}{shadow} on
\href{https://stackoverflow.com}{StackOverflow}:
\url{https://stackoverflow.com/questions/22810309/pairs-specifying-axes-limits-of-the-subpanels}
}
\examples{
data(iris)
pairs(iris[1:4], xlim = c(0, 8), ylim = c(0, 8))
pairs2(iris[1:4], xlim = c(0, 8), ylim = c(0, 8))
# xpecifying limits (now as arrays...)
# dims 1-2: panel
# dim 3: lower und upper limit
my_xlim <- array(0, dim = c(4, 4, 2))
my_xlim[, , 2] <- 8
my_ylim <- my.xlim
my_xlim[1, , 1] <- 4
pairs2(iris[1:4], xlim = my.xlim)
# careful: the following would work, but does not adjust the labels!
my_xlim[2, 3, 2] <- 6
pairs(iris[1:4], xlim = my.xlim)

}
\author{
\href{https://stackoverflow.com/users/2591234/shadow}{shadow} user
        from \href{https://stackoverflow.com}{StackOverflow}.
}
