% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishdir.R
\name{publishdir}
\alias{publishdir}
\title{Retrieve publishDir from a TOML file}
\usage{
publishdir(toml)
}
\arguments{
\item{toml}{character string naming the TOML file to parse.}
}
\value{
Returns the path of the publishDir in the input TOML file.
}
\description{
Retrieve publishDir from a TOML file
}
\author{
Marc Choisy
}
