% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Reverse rows or columns}
\usage{
reverse(x, margin)
}
\arguments{
\item{x}{a matrix or a data frame.}
}
\value{
an object of the same class as \code{x}.
}
\description{
Reverses a matrix or a data frame by row or column
}
\examples{
mtcars
reverse(mtcars, 1)
reverse(mtcars, 2)

}
