% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Report the Space Allocated for an Object}
\usage{
size(x)
}
\arguments{
\item{x}{an R object.}
}
\description{
Provides an estimate of the memory that is being used to store an R object.
}
\details{
This is a wraper around the \code{\link[utils]{object.size}} function,
returning a nicer output.
}
