% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn2latin.R
\name{vn2latin}
\alias{vn2latin}
\title{Conversion from Vietnamese characters to base Latin}
\usage{
vn2latin(x)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
The translation of x in Basic Latin.
}
\description{
\code{vn2latin} returns the base Latin equivalent of Vietnamese Unicode
characters.
}
\examples{
vn2latin("ẳ")
}
\author{
Marc Choisy
}
