% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writelines.R
\name{writelines}
\alias{writelines}
\title{Write lines to a file}
\usage{
writelines(text, file, sep = "\\n", useBytes = FALSE)
}
\arguments{
\item{text}{A character vector.}

\item{file}{The path to a file.}

\item{sep}{character string. A string to be written to the connection after
each line of text.}

\item{useBytes}{logical. See base::writeLines}
}
\description{
Write text lines to a file.
}
\details{
A wrap-up around the base::writeLines function.
}
\author{
Marc Choisy
}
